refactor(app): extract hud components
