+++ src/app/App.tsx
@@ -1,638 +1,18 @@
-import { useEffect, useMemo, useRef, useState, type FC } from 'react';
+import { useEffect, useRef, useState, type FC } from 'react';
 
 import { AttackLogPanel } from '../features/attack-log/AttackLogPanel';
 import { PlayerConfigModal } from '../features/build-config/PlayerConfigModal';
 import { useBuildConfiguration } from '../features/build-config/useBuildConfiguration';
 import { CombatLogPanel } from '../features/combat-log/CombatLogPanel';
 import {
-  CUSTOM_BOSS_ID,
   FightStateProvider,
   useFightDerivedStats,
 } from '../features/fight-state/FightStateContext';
 import { HelpModal } from '../features/help/HelpModal';
-import { formatDecimal, formatNumber, formatStopwatch } from '../utils/format';
 import { useVisualViewportCssVars } from './useVisualViewportCssVars';
-import { BossHealthBar } from './components/BossHealthBar';
-
-type StageTimelineProps = {
-  readonly stageLabel: string | null;
-  readonly stageProgress: { current: number; total: number } | null;
-  readonly onAdvance: () => void;
-  readonly onRewind: () => void;
-  readonly hasNext: boolean;
-  readonly hasPrevious: boolean;
-};
-
-const StageTimeline: FC<StageTimelineProps> = ({
-  stageLabel,
-  stageProgress,
-  onAdvance,
-  onRewind,
-  hasNext,
-  hasPrevious,
-}) => {
-  if (!stageProgress) {
-    return null;
-  }
-
-  const stageTitle = stageLabel ?? 'Current stage';
-  const stagePosition = `${stageProgress.current}/${stageProgress.total}`;
-
-  return (
-    <div
-      className="hud-timeline summary-chip summary-chip--toolbar"
-      role="group"
-      aria-label="Stage navigation"
-    >
-      <button
-        type="button"
-        className="hud-timeline__control"
-        onClick={onRewind}
-        disabled={!hasPrevious}
-      >
-        <span aria-hidden="true">‹</span>
-        <span className="sr-only">Previous stage</span>
-      </button>
-      <span className="hud-timeline__label" aria-live="polite">
-        <span className="hud-timeline__title">{stageTitle}</span>
-        <span className="hud-timeline__progress">{stagePosition}</span>
-      </span>
-      <button
-        type="button"
-        className="hud-timeline__control"
-        onClick={onAdvance}
-        disabled={!hasNext}
-      >
-        <span aria-hidden="true">›</span>
-        <span className="sr-only">Next stage</span>
-      </button>
-    </div>
-  );
-};
-
-type EncounterBrandProps = {
-  readonly encounterName: string;
-  readonly versionLabel: string | null;
-  readonly arenaLabel: string | null;
-};
-
-const EncounterBrand: FC<EncounterBrandProps> = ({
-  encounterName,
-  versionLabel,
-  arenaLabel,
-}) => (
-  <div className="hud-brand" aria-live="polite">
-    <h1 className="hud-brand__title">Hollow Knight Damage Tracker</h1>
-    <div className="hud-brand__context">
-      <span className="hud-brand__divider" aria-hidden="true">
-        ◆
-      </span>
-      <span className="hud-brand__encounter">
-        {encounterName}
-        {versionLabel ? (
-          <span className="hud-brand__version">({versionLabel})</span>
-        ) : null}
-      </span>
-      {arenaLabel ? (
-        <>
-          <span className="hud-brand__divider" aria-hidden="true">
-            ◆
-          </span>
-          <span className="hud-brand__arena">{arenaLabel}</span>
-        </>
-      ) : null}
-    </div>
-  </div>
-);
-
-type TargetScoreboardProps = {
-  readonly derived: ReturnType<typeof useFightDerivedStats>;
-  readonly stageLabel: string | null;
-  readonly stageProgress: { current: number; total: number } | null;
-  readonly onAdvanceStage: () => void;
-  readonly onRewindStage: () => void;
-  readonly hasNextStage: boolean;
-  readonly hasPreviousStage: boolean;
-};
-
-const TargetScoreboard: FC<TargetScoreboardProps> = ({
-  derived,
-  stageLabel,
-  stageProgress,
-  onAdvanceStage,
-  onRewindStage,
-  hasNextStage,
-  hasPreviousStage,
-}) => {
-  const {
-    targetHp,
-    remainingHp,
-    dps,
-    elapsedMs,
-    estimatedTimeRemainingMs,
-    averageDamage,
-    actionsPerMinute,
-    attacksLogged,
-    totalDamage,
-  } = derived;
-  const metrics = useMemo(
-    () => [
-      { id: 'elapsed', label: 'Elapsed', value: formatStopwatch(elapsedMs) },
-      {
-        id: 'estimated-remaining',
-        label: 'Est. Remaining',
-        value: formatStopwatch(estimatedTimeRemainingMs),
-      },
-      {
-        id: 'dps',
-        label: 'DPS',
-        value: formatDecimal(dps),
-        sublabel: typeof totalDamage === 'number' ? formatNumber(totalDamage) : undefined,
-      },
-      {
-        id: 'average-damage',
-        label: 'Avg Dmg',
-        value: formatDecimal(averageDamage),
-      },
-      {
-        id: 'actions-per-minute',
-        label: 'APM',
-        value: formatDecimal(actionsPerMinute),
-        sublabel:
-          typeof attacksLogged === 'number' ? formatNumber(attacksLogged) : undefined,
-      },
-    ],
-    [
-      actionsPerMinute,
-      attacksLogged,
-      averageDamage,
-      dps,
-      elapsedMs,
-      estimatedTimeRemainingMs,
-      totalDamage,
-    ],
-  );
-
-  return (
-    <section className="hud-scoreboard" aria-label="Encounter scoreboard">
-      <div className="hud-scoreboard__status">
-        <BossHealthBar
-          className="hud-health summary-chip summary-chip--accent"
-          role="group"
-          aria-label="Boss HP"
-          label="HP"
-          current={remainingHp}
-          total={targetHp}
-          progressbarAriaLabel="Boss HP"
-          valueLabel={`${formatNumber(remainingHp)} / ${formatNumber(targetHp)}`}
-        />
-        <StageTimeline
-          stageLabel={stageLabel}
-          stageProgress={stageProgress}
-          onAdvance={onAdvanceStage}
-          onRewind={onRewindStage}
-          hasNext={hasNextStage}
-          hasPrevious={hasPreviousStage}
-        />
-      </div>
-      <dl className="hud-metrics">
-        {metrics.map((metric) => (
-          <div key={metric.id} className="hud-metrics__item" data-metric-id={metric.id}>
-            <dt className="hud-metrics__label">{metric.label}</dt>
-            <dd className="hud-metrics__value">
-              <span className="hud-metrics__value-primary">{metric.value}</span>
-              {metric.sublabel ? (
-                <span className="hud-metrics__sublabel">{metric.sublabel}</span>
-              ) : null}
-            </dd>
-          </div>
-        ))}
-      </dl>
-    </section>
-  );
-};
-
-type MobilePinnedHudProps = {
-  readonly derived: ReturnType<typeof useFightDerivedStats>;
-  readonly encounterName: string;
-};
-
-const MobilePinnedHud: FC<MobilePinnedHudProps> = ({ derived, encounterName }) => {
-  const { targetHp, remainingHp } = derived;
-  return (
-    <div className="mobile-hud-sentinel">
-      <div className="mobile-hud" role="group" aria-label="Boss status">
-        <span className="mobile-hud__title" aria-live="polite">
-          {encounterName}
-        </span>
-        <BossHealthBar
-          className="mobile-hud__health"
-          role="group"
-          aria-label="Boss HP"
-          current={remainingHp}
-          total={targetHp}
-          progressbarAriaLabel="Boss HP"
-          trackClassName="mobile-hud__track"
-          valueLabel={`${formatNumber(remainingHp)} / ${formatNumber(targetHp)}`}
-          valueClassName="mobile-hud__value"
-        />
-      </div>
-    </div>
-  );
-};
-
-type HeaderBarProps = {
-  readonly derived: ReturnType<typeof useFightDerivedStats>;
-  readonly encounterName: string;
-  readonly versionLabel: string | null;
-  readonly arenaLabel: string | null;
-  readonly onToggleSetup: () => void;
-  readonly onOpenLoadout: () => void;
-  readonly onOpenHelp: () => void;
-  readonly isSetupOpen: boolean;
-  readonly stageLabel: string | null;
-  readonly stageProgress: { current: number; total: number } | null;
-  readonly onAdvanceStage: () => void;
-  readonly onRewindStage: () => void;
-  readonly hasNextStage: boolean;
-  readonly hasPreviousStage: boolean;
-};
-
-const HeaderBar: FC<HeaderBarProps> = ({
-  derived,
-  encounterName,
-  versionLabel,
-  arenaLabel,
-  onToggleSetup,
-  onOpenLoadout,
-  onOpenHelp,
-  isSetupOpen,
-  stageLabel,
-  stageProgress,
-  onAdvanceStage,
-  onRewindStage,
-  hasNextStage,
-  hasPreviousStage,
-}) => (
-  <header className="encounter-hud app-navbar" role="banner">
-    <div className="encounter-hud__primary">
-      <EncounterBrand
-        encounterName={encounterName}
-        versionLabel={versionLabel}
-        arenaLabel={arenaLabel}
-      />
-      <div className="hud-actions">
-        <button
-          type="button"
-          className="hud-actions__button"
-          onClick={onToggleSetup}
-          aria-expanded={isSetupOpen}
-          aria-controls="encounter-setup"
-        >
-          <span aria-hidden="true">⚙️</span>
-          <span className="hud-actions__label">Change encounter</span>
-        </button>
-        <button type="button" className="hud-actions__button" onClick={onOpenLoadout}>
-          <span aria-hidden="true">👤</span>
-          <span className="hud-actions__label">Player loadout</span>
-        </button>
-        <button type="button" className="hud-actions__button" onClick={onOpenHelp}>
-          <span aria-hidden="true">❓</span>
-          <span className="hud-actions__label">Help</span>
-        </button>
-      </div>
-    </div>
-    <TargetScoreboard
-      derived={derived}
-      stageLabel={stageLabel}
-      stageProgress={stageProgress}
-      onAdvanceStage={onAdvanceStage}
-      onRewindStage={onRewindStage}
-      hasNextStage={hasNextStage}
-      hasPreviousStage={hasPreviousStage}
-    />
-  </header>
-);
-
-type TargetSelectorProps = {
-  readonly bosses: ReturnType<typeof useBuildConfiguration>['bosses'];
-  readonly bossSelectValue: string;
-  readonly onBossChange: (value: string) => void;
-  readonly selectedBoss: ReturnType<typeof useBuildConfiguration>['selectedBoss'];
-  readonly selectedBossId: string | null;
-  readonly onBossVersionChange: (value: string) => void;
-  readonly selectedTarget: ReturnType<typeof useBuildConfiguration>['selectedTarget'];
-  readonly selectedVersion: ReturnType<typeof useBuildConfiguration>['selectedVersion'];
-  readonly customTargetHp: number;
-  readonly onCustomHpChange: (value: string) => void;
-};
-
-const TargetSelector: FC<TargetSelectorProps> = ({
-  bosses,
-  bossSelectValue,
-  onBossChange,
-  selectedBoss,
-  selectedBossId,
-  onBossVersionChange,
-  selectedTarget,
-  selectedVersion,
-  customTargetHp,
-  onCustomHpChange,
-}) => {
-  const [isOptionsOpen, setOptionsOpen] = useState(false);
-  const [customHpDraft, setCustomHpDraft] = useState(() => customTargetHp.toString());
-
-  useEffect(() => {
-    if (selectedBossId === CUSTOM_BOSS_ID) {
-      setOptionsOpen(true);
-    }
-  }, [selectedBossId]);
-
-  useEffect(() => {
-    setCustomHpDraft(customTargetHp.toString());
-  }, [customTargetHp]);
-
-  const handleCustomHpDraftChange = (value: string) => {
-    const sanitized = value.replace(/[^0-9]/g, '');
-    setCustomHpDraft(sanitized);
-    if (sanitized !== '') {
-      onCustomHpChange(sanitized);
-    }
-  };
-
-  return (
-    <section className="target-selector" aria-labelledby="target-selector-heading">
-      <div className="target-selector__header">
-        <h3 id="target-selector-heading">Boss target</h3>
-        <button
-          type="button"
-          className="target-selector__options-toggle"
-          onClick={() => setOptionsOpen((open) => !open)}
-          aria-expanded={isOptionsOpen}
-          aria-controls="target-selector-options"
-        >
-          ⚙️
-          <span className="sr-only">Toggle advanced target options</span>
-        </button>
-      </div>
-      <div className="segmented-control" role="radiogroup" aria-label="Boss target">
-        {bosses.map((boss) => {
-          const isSelected = bossSelectValue === boss.id;
-          return (
-            <button
-              key={boss.id}
-              type="button"
-              className="segmented-control__option"
-              data-selected={isSelected}
-              onClick={() => onBossChange(boss.id)}
-              role="radio"
-              aria-checked={isSelected}
-            >
-              {boss.name}
-            </button>
-          );
-        })}
-        <button
-          type="button"
-          className="segmented-control__option"
-          data-selected={bossSelectValue === CUSTOM_BOSS_ID}
-          onClick={() => onBossChange(CUSTOM_BOSS_ID)}
-          role="radio"
-          aria-checked={bossSelectValue === CUSTOM_BOSS_ID}
-        >
-          Custom
-        </button>
-      </div>
-      <div
-        id="target-selector-options"
-        className="target-selector__tray"
-        hidden={!isOptionsOpen}
-      >
-        {selectedBoss && selectedBossId !== CUSTOM_BOSS_ID && selectedTarget ? (
-          <label className="target-selector__field">
-            <span className="target-selector__field-label">Boss version</span>
-            <select
-              value={selectedTarget.id}
-              onChange={(event) => onBossVersionChange(event.target.value)}
-            >
-              {selectedBoss.versions.map((version) => (
-                <option key={version.targetId} value={version.targetId}>
-                  {`${version.title} • ${version.hp.toLocaleString()} HP`}
-                </option>
-              ))}
-            </select>
-          </label>
-        ) : null}
-
-        {selectedBossId === CUSTOM_BOSS_ID ? (
-          <label className="target-selector__field" htmlFor="custom-target-hp">
-            <span className="target-selector__field-label">Custom target HP</span>
-            <input
-              id="custom-target-hp"
-              type="text"
-              inputMode="numeric"
-              pattern="[0-9]*"
-              value={customHpDraft}
-              onChange={(event) => handleCustomHpDraftChange(event.target.value)}
-              onBlur={() => {
-                if (customHpDraft === '') {
-                  setCustomHpDraft(customTargetHp.toString());
-                }
-              }}
-            />
-          </label>
-        ) : null}
-
-        {selectedTarget && selectedVersion ? (
-          <div className="target-selector__summary summary-chip">
-            <span className="target-selector__summary-title">Active target</span>
-            <span className="target-selector__summary-value">
-              {selectedTarget.bossName}
-            </span>
-            <span className="target-selector__summary-meta">{selectedVersion.title}</span>
-          </div>
-        ) : null}
-      </div>
-    </section>
-  );
-};
-
-type SequenceSelectorProps = {
-  readonly bossSequences: ReturnType<typeof useBuildConfiguration>['bossSequences'];
-  readonly sequenceSelectValue: string;
-  readonly onSequenceChange: (value: string) => void;
-  readonly sequenceEntries: ReturnType<typeof useBuildConfiguration>['sequenceEntries'];
-  readonly cappedSequenceIndex: number;
-  readonly onStageSelect: (index: number) => void;
-};
-
-const SequenceSelector: FC<SequenceSelectorProps> = ({
-  bossSequences,
-  sequenceSelectValue,
-  onSequenceChange,
-  sequenceEntries,
-  cappedSequenceIndex,
-  onStageSelect,
-}) => (
-  <section className="sequence-selector" aria-labelledby="sequence-selector-heading">
-    <div className="sequence-selector__header">
-      <h3 id="sequence-selector-heading">Encounter stage</h3>
-    </div>
-    <label className="sequence-selector__field">
-      <span className="sequence-selector__field-label">Mode</span>
-      <select
-        value={sequenceSelectValue}
-        onChange={(event) => onSequenceChange(event.target.value)}
-      >
-        <option value="">Single target practice</option>
-        {bossSequences.map((sequence) => (
-          <option key={sequence.id} value={sequence.id}>
-            {sequence.name}
-          </option>
-        ))}
-      </select>
-    </label>
-    {sequenceEntries.length > 0 ? (
-      <ol className="sequence-selector__stages">
-        {sequenceEntries.map((entry, index) => {
-          const isCurrent = index === cappedSequenceIndex;
-          return (
-            <li key={entry.id}>
-              <button
-                type="button"
-                className="sequence-selector__stage"
-                onClick={() => onStageSelect(index)}
-                aria-current={isCurrent ? 'true' : undefined}
-              >
-                <span className="sequence-selector__stage-index">
-                  {String(index + 1).padStart(2, '0')}
-                </span>
-                <span className="sequence-selector__stage-name">
-                  {entry.target.bossName}
-                </span>
-              </button>
-            </li>
-          );
-        })}
-      </ol>
-    ) : (
-      <p className="sequence-selector__empty" aria-live="polite">
-        Select a Godhome sequence to practice multi-fight runs.
-      </p>
-    )}
-  </section>
-);
-
-type EncounterSetupPanelProps = {
-  readonly isOpen: boolean;
-  readonly bosses: ReturnType<typeof useBuildConfiguration>['bosses'];
-  readonly bossSelectValue: string;
-  readonly onBossChange: (value: string) => void;
-  readonly selectedBoss: ReturnType<typeof useBuildConfiguration>['selectedBoss'];
-  readonly selectedBossId: string | null;
-  readonly onBossVersionChange: (value: string) => void;
-  readonly selectedTarget: ReturnType<typeof useBuildConfiguration>['selectedTarget'];
-  readonly selectedVersion: ReturnType<typeof useBuildConfiguration>['selectedVersion'];
-  readonly customTargetHp: number;
-  readonly onCustomHpChange: (value: string) => void;
-  readonly bossSequences: ReturnType<typeof useBuildConfiguration>['bossSequences'];
-  readonly sequenceSelectValue: string;
-  readonly onSequenceChange: (value: string) => void;
-  readonly sequenceEntries: ReturnType<typeof useBuildConfiguration>['sequenceEntries'];
-  readonly cappedSequenceIndex: number;
-  readonly onStageSelect: (index: number) => void;
-  readonly activeSequence: ReturnType<typeof useBuildConfiguration>['activeSequence'];
-  readonly sequenceConditionValues: ReturnType<
-    typeof useBuildConfiguration
-  >['sequenceConditionValues'];
-  readonly onConditionToggle: (conditionId: string, enabled: boolean) => void;
-};
-
-const EncounterSetupPanel: FC<EncounterSetupPanelProps> = ({
-  isOpen,
-  bosses,
-  bossSelectValue,
-  onBossChange,
-  selectedBoss,
-  selectedBossId,
-  onBossVersionChange,
-  selectedTarget,
-  selectedVersion,
-  customTargetHp,
-  onCustomHpChange,
-  bossSequences,
-  sequenceSelectValue,
-  onSequenceChange,
-  sequenceEntries,
-  cappedSequenceIndex,
-  onStageSelect,
-  activeSequence,
-  sequenceConditionValues,
-  onConditionToggle,
-}) => (
-  <section
-    id="encounter-setup"
-    className="encounter-setup"
-    aria-label="Encounter setup"
-    hidden={!isOpen}
-  >
-    <div className="encounter-setup__grid">
-      <TargetSelector
-        bosses={bosses}
-        bossSelectValue={bossSelectValue}
-        onBossChange={onBossChange}
-        selectedBoss={selectedBoss}
-        selectedBossId={selectedBossId}
-        onBossVersionChange={onBossVersionChange}
-        selectedTarget={selectedTarget}
-        selectedVersion={selectedVersion}
-        customTargetHp={customTargetHp}
-        onCustomHpChange={onCustomHpChange}
-      />
-      <SequenceSelector
-        bossSequences={bossSequences}
-        sequenceSelectValue={sequenceSelectValue}
-        onSequenceChange={onSequenceChange}
-        sequenceEntries={sequenceEntries}
-        cappedSequenceIndex={cappedSequenceIndex}
-        onStageSelect={onStageSelect}
-      />
-    </div>
-    {activeSequence && activeSequence.conditions.length > 0 ? (
-      <section
-        className="sequence-conditions"
-        aria-label="Sequence conditions"
-        role="group"
-      >
-        <h4 className="sequence-conditions__title">Sequence conditions</h4>
-        <div className="sequence-conditions__grid">
-          {activeSequence.conditions.map((condition) => {
-            const isEnabled = sequenceConditionValues[condition.id];
-            return (
-              <label key={condition.id} className="sequence-conditions__option">
-                <input
-                  type="checkbox"
-                  checked={isEnabled}
-                  onChange={(event) =>
-                    onConditionToggle(condition.id, event.target.checked)
-                  }
-                />
-                <span>
-                  <span className="sequence-conditions__label">{condition.label}</span>
-                  {condition.description ? (
-                    <span className="sequence-conditions__description">
-                      {condition.description}
-                    </span>
-                  ) : null}
-                </span>
-              </label>
-            );
-          })}
-        </div>
-      </section>
-    ) : null}
-  </section>
-);
+import { EncounterSetupPanel } from './components/EncounterSetupPanel';
+import { HeaderBar } from './components/HeaderBar';
+import { MobilePinnedHud } from './components/MobilePinnedHud';
 
 const AppContent: FC = () => {
   useVisualViewportCssVars();
